기기 on/off 까지 완성
